requirements.txt
lxml==4.9.2
lzstring==1.0.4
PyExecJS==1.5.1
Requests==2.32.3
tqdm==4.66.1
